Add DefaultBrowserPromptingState2 preference
Through some experimentation, determined that the DefaultBrowserPromptingState2 preference key controls whether or not a prompt is created following Safari quit to offer keeping the current default browser (if not Safari) or changing it to Safari.  Enabling this preference with an integer value of 1 successfully suppresses this prompt.

As a result of this addition, also shifted the formatting left a tab.