Fix Compressor Manifest (#251)
* Add FFCheckedUpgrade

With just a dictionary with `{{name}}` and `{{value}}`, nothing appears in the GUI for the payload.

With at least one concrete preference added, it displays correctly.  Preference copied from Motion manifest.  Since it is unclear as to whether this preference actually applies to Compressor, hide it from the GUI with `pfm_hidden`.  Confirmed that this does not add it to the GUI or XML, but that original dictionary and contained prefs appear now.

* Update last_modified date