Update Preferences dictionary subkey formatting
In Firefox 81, the format strucuture for preference subkeys within the larger 'Preferences' dictionary was changed such that each subkey is now it's own dictionary.  This manifest has been updated to reflect this and the newly introduced 'Status' subkey was added for each.

Given the nature of this change, this would prevent admins from deploying the 'Preferences' dictionary to Macs for prior versions.

As part of these changes, the duplicate pfm_titles were also removed.

New preferences were included, but their value types have not been defined in either v2.3 or v2.4 of the Firefox policy releases