Adjust XML comment escaping for Nudge plist
The link to the Nudge wiki in the commented-out `attemptToFetchMajorUpgrade` setting is preventing the plist from being parsed by Python's `plistlib`. This is probably because the `---` in the URL is being parsed as the end of a comment.

Adding backslashes to escape the hyphens successfully allows the plist to be parsed, while still allowing the URL to be usable without manual changes.

Another alternative would be to encode the hyphens as `&#45;&#45;&#45;` but I thought the backslashes would be more readable.